requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==10.1.0
openai-whisper==20231117
yt-dlp==2023.11.16
transformers==4.35.2
torch==2.4.1
sentencepiece==0.2.0
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1